First Upload and Quad Method
Creation of the Repo and the upload of a quadding method with working
examples.

The two examples included allow you to test the main quad check method
then a second test for using that quad method to create a fast ordered
list builder.

With these two elements It should be relatively straight forward to add
a removal method for items in the quad; this'll round out the examples
to make full use of what they can do,